Netcdf support (#39)
* Adding netcdf support.
Refactored gribfile to pull out the statistics calculations and have separate file readers for different file types.
The statistics calculations are now more flexible in terms of stat names and dimensions to average over as they will be used by a wider variety of data formats.
Made the new stats class inherit ConfigMixin so it can be added to yaml configs.

* Change stats tests from grib to netcdf so they can run when grib is unavailable. Add tests to improve coverage.

* Split NetcdfFileReader into separate file. Add some doc strings.